requirements.txt
streamlit>=1.24,<2
python-dotenv>=1.0,<2
Pillow>=10.0,<11
requests>=2.30,<3
